Extract common date/time code to lib/datetime
